+++ src/frontend/components/CartItems/CartItem.tsx
@@ -1,0 +1,37 @@
+import Link from 'next/link';
+import { Product } from '../../protos/demo';
+import ProductPrice from '../ProductPrice';
+import * as S from './CartItems.styled';
+
+interface IProps {
+  product: Product;
+  quantity: number;
+}
+
+const CartItem = ({
+  product: { id, name, picture, priceUsd = { units: 0, nanos: 0, currencyCode: 'USD' } },
+  quantity,
+}: IProps) => {
+  return (
+    <S.CartItem>
+      <Link href={`/product/${id}`}>
+        <S.NameContainer>
+          <S.CartItemImage alt={name} src={picture} />
+          <p>{name}</p>
+        </S.NameContainer>
+      </Link>
+      <S.CartItemDetails>
+        <p>{quantity}</p>
+      </S.CartItemDetails>
+      <S.CartItemDetails>
+        <S.PriceContainer>
+          <p>
+            <ProductPrice price={priceUsd} />
+          </p>
+        </S.PriceContainer>
+      </S.CartItemDetails>
+    </S.CartItem>
+  );
+};
+
+export default CartItem;